Split frontend
