+++ save-cosv-frontend/src/main/kotlin/com/saveourtool/save/cosv/frontend/components/views/vuln/CreateVulnerabilityView.kt
@@ -463,9 +463,6 @@ private fun ChildrenBuilder.requiredTitle(title: String, htmlForId: String? = nu
     }
 }
 
-@Suppress(
-    "MagicNumber",
-)
 private fun getColor(value: Float): String = when (value) {
     0f -> Colors.NONE.value
     in 0.1f..3.9f -> Colors.SUCCESS.value
