+++ save-cosv-frontend/src/main/kotlin/com/saveourtool/save/cosv/frontend/components/views/vuln/UploadVulnerabilityView.kt
@@ -2,7 +2,7 @@
 
 package com.saveourtool.save.cosv.frontend.components.views.vuln
 
-import com.saveourtool.save.frontend.common.components.basic.fileuploader.cosvFileManagerComponent
+import com.saveourtool.save.cosv.frontend.components.views.vuln.component.cosvFileManagerComponent
 import com.saveourtool.save.frontend.common.components.modal.MAX_Z_INDEX
 import com.saveourtool.save.frontend.common.externals.i18next.useTranslation
 import com.saveourtool.save.frontend.common.utils.Style
