+++ save-cosv-frontend/src/main/kotlin/com/saveourtool/save/cosv/frontend/components/views/vuln/VulnerabilityGeneralInfoProps.kt
@@ -43,7 +43,7 @@ import kotlinx.datetime.TimeZone
 /**
  * [FC] that is used to display some general vulnerability information
  */
-@Suppress("EMPTY_BLOCK_STRUCTURE_ERROR", "MAGIC_NUMBER", "MagicNumber")
+@Suppress("EMPTY_BLOCK_STRUCTURE_ERROR", "MAGIC_NUMBER")
 val vulnerabilityGeneralInfoProps: FC<VulnerabilityGeneralInfoProps> = FC { props ->
     val (t) = useTranslation("vulnerability")
     val (cachedVulnerability, setCachedVulnerability) = useState<VulnerabilityExt>()
