+++ save-cosv-frontend/src/main/kotlin/com/saveourtool/save/cosv/frontend/components/views/vuln/VulnerabilityHeader.kt
@@ -144,7 +144,6 @@ internal val headerMenu: FC<HeaderMenuProps> = FC { props ->
         }
     }
 
-    @Suppress("MagicNumber")
     displayModal(
         rejectVulnerabilityWindowOpenness.isOpen(),
         if (buttonType == ButtonType.REJECT) "Reject vulnerability".t() else "Ready for review".t(),
