+++ save-cosv-frontend/src/main/kotlin/com/saveourtool/save/cosv/frontend/components/views/vuln/VulnerabilityInfoTab.kt
@@ -6,12 +6,12 @@
 
 package com.saveourtool.save.cosv.frontend.components.views.vuln
 
+import com.saveourtool.save.cosv.frontend.components.views.vuln.component.timelineComponent
 import com.saveourtool.save.entities.cosv.VulnerabilityExt
 import com.saveourtool.save.entities.vulnerability.VulnerabilityDateDto
 import com.saveourtool.save.entities.vulnerability.VulnerabilityDateType
 import com.saveourtool.save.entities.vulnerability.VulnerabilityProjectType
 import com.saveourtool.save.frontend.common.components.basic.carousel
-import com.saveourtool.save.frontend.common.components.basic.timelineComponent
 import com.saveourtool.save.frontend.common.components.modal.displayModal
 import com.saveourtool.save.frontend.common.components.modal.mediumTransparentModalStyle
 import com.saveourtool.save.frontend.common.components.tables.TableProps
@@ -316,7 +316,7 @@ val vulnerabilityInfoTab: FC<VulnerabilityInfoTabProps> = FC { props ->
     @Suppress("COMPLEX_EXPRESSION")
     renderProjects(
         "Commits with fix".t(),
-        "affectedCarousel",
+        "commitsCarousel",
         VulnerabilityProjectType.COMMIT,
         vulnerabilityProjects,
         commitFixTable,
