--- save-frontend-common/src/main/kotlin/com/saveourtool/save/frontend/common/components/basic/codeeditor/FileType.kt
@@ -1,52 +1,0 @@
-package com.saveourtool.save.frontend.common.components.basic.codeeditor
-
-import com.saveourtool.save.utils.Languages
-
-/**
- * @property fileName displayed name
- * @property urlPart part of url for uploading\downloading text
- * @property editorMode highlight mode that should be enabled, if null, mode can be chosen using selector
- */
-enum class FileType(
-    val fileName: String,
-    val urlPart: String,
-    val editorMode: Languages?,
-) {
-    SAVE_TOML("save.toml", "test", Languages.TOML),
-    SETUP_SH("setup.sh", "file", Languages.SHELL),
-    TEST("test", "test", null),
-    ;
-    companion object {
-        /**
-         * @param fileType [FileType]
-         * @param optionForCode will be returned if [fileType] is CODE
-         * @param optionForSaveToml will be returned if [fileType] is SAVE_TOML
-         * @param optionForSetupSh will be returned if [fileType] is SETUP_SH
-         * @return option corresponding to [fileType]
-         */
-        fun <T> getTypedOption(
-            fileType: FileType,
-            optionForCode: T,
-            optionForSaveToml: T,
-            optionForSetupSh: T,
-        ) = when (fileType) {
-            TEST -> optionForCode
-            SAVE_TOML -> optionForSaveToml
-            SETUP_SH -> optionForSetupSh
-        }
-
-        /**
-         * @param fileType [FileType]
-         * @param optionForCode will be returned if [fileType] is CODE
-         * @param optionForSaveToml will be returned if [fileType] is SAVE_TOML
-         * @param optionForSetupSh will be returned if [fileType] is SETUP_SH
-         * @return option corresponding to [fileType] or null if [fileType] is null
-         */
-        fun <T> getTypedOption(
-            fileType: FileType?,
-            optionForCode: T,
-            optionForSaveToml: T,
-            optionForSetupSh: T,
-        ) = fileType?.let { getTypedOption(it, optionForCode, optionForSaveToml, optionForSetupSh) }
-    }
-}
